+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/ProcessValidation.scala
@@ -59,7 +59,7 @@ class ProcessValidation(modelData: ProcessingTypeDataProvider[ModelData],
   def withAdditionalPropertiesConfig(additionalPropertiesConfig: ProcessingTypeDataProvider[Map[String, AdditionalPropertyConfig]]) =
     new ProcessValidation(modelData, additionalPropertiesConfig, additionalValidators, subprocessResolver, None)
 
-  def validate(displayable: DisplayableProcess, category: Category): ValidationResult = {
+  def validate(displayable: DisplayableProcess): ValidationResult = {
     val uiValidationResult = uiValidation(displayable)
 
     //there is no point in further validations if ui process structure is invalid
@@ -67,7 +67,7 @@ class ProcessValidation(modelData: ProcessingTypeDataProvider[ModelData],
     if (uiValidationResult.saveAllowed) {
       val canonical = ProcessConverter.fromDisplayable(displayable)
       uiValidationResult
-        .add(processingTypeValidationWithTypingInfo(canonical, displayable.processingType, category))
+        .add(processingTypeValidationWithTypingInfo(canonical, displayable.processingType, displayable.category))
     } else {
       uiValidationResult
     }