+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 ------------------------
 * [#3986](https://github.com/TouK/nussknacker/pull/3986) Updated sttp 2.2.9 -> 3.8.11
 * [#3979](https://github.com/TouK/nussknacker/pull/3979) Updated Flink 1.16.0 -> 1.16.1
+* [#4019](https://github.com/TouK/nussknacker/pull/4019) Make displayable process category required
 
 1.8.0 (Not released yet)
 ------------------------